Hardcode comment data on DataServlet and return as JSON response
Hardcode and store dummy comment data in ArrayList in DataServlet and convert to JSON response to client requests to /data URL. This is to check that JSON server responses work as they should.
This data will be queried by front-end and displayed in future. Hardcoded comment data will also be eventually be replaced with real user data in later stages of project.