Fetch hardcoded, dummy comment data from server and display on webpage (#5)
* Hardcode comment data on DataServlet and return as JSON response

Hardcode and store dummy comment data in ArrayList in DataServlet and convert to JSON response to client requests to /data URL. This is to check that JSON server responses work as they should.
This data will be queried by front-end and displayed in future. Hardcoded comment data will also be eventually be replaced with real user data in later stages of project.

* Add addCommentData function retrieving dummy comment data from /data URL  and display the data in front-end

addCommentData is fired onclick of button in index.html, and retrieves comment data in JSON format, which is parsed and outputted to index.html page.
This is done to verify that JSON server responses can be queried, received, parsed and displayed successfully.