Add comment form element to index.html
Add <form> element for user to submit comments from the index.html page, and add CSS styling to this form. Form includes fields for the comment text, and commenter name.
Submit button on form currently linked to a doPost function in the servlet at /data URL that just returns. In future, this doPost will store the data collected from the form on the server.