Implement doPost function to get and store comment data
doPost in DataServlet will retrieve the comment text and author name data from the comment <form> in index.html and store the strings in the ArrayList.
Note, this is to check that the DataServlet can handle HTTP POST requests, and can store data appropriately, as per step/walkthroughs/week-3-server/step-4-post-walkthrough.md.