Modify addCommentData (now fetchComments) to retrieve and display user comments
addCommentData in script.js  was renamed to fetchComments for brevity.
function will now fetch the comments from servlet at /data URL (comment data which is submitted there via the comment form on index.html), and display on front-end.
note that comment data is currently not persistent between sessions, but will be made persistent in future modifications.