Add comment form on portfolio home page (#6)
* Remove hardcoded comment data from servlet

Hardcoded data will be replaced with real user data collected from front-end comment form in future.

* Add comment form element to index.html

Add <form> element for user to submit comments from the index.html page, and add CSS styling to this form. Form includes fields for the comment text, and commenter name.
Submit button on form currently linked to a doPost function in the servlet at /data URL that just returns. In future, this doPost will store the data collected from the form on the server.

* Implement doPost function to get and store comment data

doPost in DataServlet will retrieve the comment text and author name data from the comment <form> in index.html and store the strings in the ArrayList.
Note, this is to check that the DataServlet can handle HTTP POST requests, and can store data appropriately, as per step/walkthroughs/week-3-server/step-4-post-walkthrough.md.

* Modify addCommentData (now fetchComments) to retrieve and display user comments

addCommentData in script.js  was renamed to fetchComments for brevity.
function will now fetch the comments from servlet at /data URL (comment data which is submitted there via the comment form on index.html), and display on front-end.
note that comment data is currently not persistent between sessions, but will be made persistent in future modifications.