Retrieve url of  uploaded image from BlobstoreService and store in datastore
When user submits form, the form will send inputted data to the URL returned by the BlobstoreUploadUrlServlet, where the image is processed and stored.
The BlobstoreUploadUrlServlet will then redirect to the DataServlet at '/data' endpoint where the image url is then retreived from the BlobstoreService and stored in datastore as part of the Comment entity.
Also added encoding type in <form> element in index.html - necessary for successful storage of the image.