Minor refactoring to comments functionality (#10)
* Add project ID to pom.xml

* Make "Comment" field in comment form compulsory

Users could submit the comment form with no text in either of the comment text or comment author fields. Fixed this to make the comment field compulsory and the author name field optional - where author name is not provided, 'Anonymous' will be displayed in place of the name.
Added '*' in UI next to comment field to indicate that the field is compulsory.