Support image upload and storing using Blobstore (#11)
* Rename comment form div ID

Rename the div containing comment form (wrapper div for styling purposes): 'comment-form' --> 'comment-form-div'.
Add id='somment-form' to the form element representing the comment form.
This is to prepare for mapping the blobstore upload url to the form action (need to be able to identify the form by an id).

* Add image upload field in comment form and fix style

Fix the spacing between form fields.
Blobstore and other backend functionality will be linked to this form in future.

* Add servlet that returns the blobstore upload url

* Fetch blobstore upload url from servlet on page load and add to form

Call getBlobstoreUrl() onload to fetch the blobstore upload url from the BlobstoreUploadUrlServlet, and add as the 'action' of the comment-form.

* Retrieve url of  uploaded image from BlobstoreService and store in datastore

When user submits form, the form will send inputted data to the URL returned by the BlobstoreUploadUrlServlet, where the image is processed and stored.
The BlobstoreUploadUrlServlet will then redirect to the DataServlet at '/data' endpoint where the image url is then retreived from the BlobstoreService and stored in datastore as part of the Comment entity.
Also added encoding type in <form> element in index.html - necessary for successful storage of the image.

* Update class header comment to describe class function

* Send uploaded image from backend to frontend and display

Add uploaded image url to the JSON object returned on GET from DataServlet, and display the image with associated comment. Images are also linked to their respective urls.

* Fix style of comments to fit in uploaded images

Edit script.js to add <div>s  to wrap around each of comment text and comment image elements when generating the comment output. Better divides each comment into separate sections for CSS styling and layout on the page.

* Delete stored images when comment is deleted

Add backend support to delete Blob storing the image when the associated comment are deleted.
Previously, all comment entities were deleted but the actual images would still be stored somewhere using BlobStore.
Now, all images are deleted along with their BlobInfo metadata.

* remove unnecessary blank lines

* Refactor displaying of comment images to reduce repeated code

Move the shared section of the HTML string representing a comment to outside of the if-block.