Add servlet, and form on front-end for user to set nickname
NicknameServlet doGet will prompt the user to login (if not logged in) and will redirect the user to the createUserProfile.html to input a nickname if they do not have a nickname, before redirecting back to index.html.
Form on createUserProfile.html allows user to submit a nickname, and is linked to the doPost in the NicknameServlet, which will store the chosen nickname under a UserInfo entity in Datastore.