Hide email unless requested by commenter
Add checkbox to comment form for user to choose whether to reveal their email in their comment.
Add backend support in DataServer to store this input field in datastore, and return the input in this field in the response JSON.
Also modify the comments displaying to support this (only display email if user has specified so).