Add image blobkey to response from DataServlet and script.js to fetch images using blobkey
DataServlet no longer uses ImageService to generate a URL for the image, and instead returns the BlobKey string of the image in its response.
script.js will pass this blobkey to fetchCommentImage to fetch the image, and generate a URL for it. This URL will then be used to display the image with the associated comment.
In this way, the ImageService dependency that does not work in production is removed.