Move getUserNickname function to new Nickname class
getUserNickname is now a static function in Nickname, a class containing helper functions to facilitate user nicknames.
This function was previously duplicated in both DataServlet and NicknameServlet, which is repeated code and not ideal practice.
The function was not made into a static function in the servlet classes, for it was not considered to be closely related to either servlets' key purpose and operations.