Add support for meeting optional attendees
Consider optional attendees in the algorithm to find meeting times.

NOTE: Optional attendees will be ignored completely if their lack of availability causes there to be no meeting times available (in which case only mandatory attendees' schedules will be considered.)

Algorithm is as follows:
1. Going through all the events, add events that involve meeting attendees (both optional and mandatory) to an ArrayList, eventList.
2. Sort eventList according to event start times
3. Merge adjacent events if they overlap
4. Going through the merged events list, add the gaps between events that are longer or  equal to meeting duration to output
5. If there are no meeting times, remove optional attendees, and repeat steps 1-4.

To facilitate the repeat of steps of 1-4, the original query function was also broken up into smaller functions to avoid repeated code.

Also added tests to verify the correctness of the support for optional attendees. All tests pass.