Display user nickname instead of name submitted with form or their email address (#16)
* Remove "Name" field from frontend and display logged in user nickname instead

Name field is redundant since only logged in users can access comment form, and all logged in users will have set a nickname.
Instead of fetching commentAuthor from the Name field, fetch the user nickname from the asociated UserInfo entity, and store the user nickname in the commentAuthor field, and display this where the input from the Name field would normally be displayed.

* Hide email unless requested by commenter

Add checkbox to comment form for user to choose whether to reveal their email in their comment.
Add backend support in DataServer to store this input field in datastore, and return the input in this field in the response JSON.
Also modify the comments displaying to support this (only display email if user has specified so).

* Fix style: add more space between each input field in comment form

* Move getUserNickname function to new Nickname class

getUserNickname is now a static function in Nickname, a class containing helper functions to facilitate user nicknames.
This function was previously duplicated in both DataServlet and NicknameServlet, which is repeated code and not ideal practice.
The function was not made into a static function in the servlet classes, for it was not considered to be closely related to either servlets' key purpose and operations.

* Only send user email in response if user selected to reveal email

For better privacy.
Previously email was included in reponse regardless of the selection, and was just not displayed. However the email could be viewed in Dev Tools, which undos the purpose of the option to show/hide email, which was to provide additional privacy.

* Change showEmail field to be boolean instead of string type

showEmail field is stored in Datastore as a boolean and returned by DataServlet as a boolean in the JSON response as well.
This makes more sense as showEmail is inherently a boolean flag.