Handle blob-key exceptional cases
When invalid blobkey is submitted to the /serve-images server, the server will respond with a 404 (image not found), which is handled in the front end by displaying "image not found" to the user, and console logging the error.
When no blobkey is provided, the server returns will return with null.