Remove ImageService dependency and serve image blobs direct from BlobStore (#17)
* New ImageBlobsServlet serving image blobs directly from BlobStore

Images will be retrieved using a GET request with a blobkey via this servlet rather than using ImageService.

* Add image blobkey to response from DataServlet and script.js to fetch images using blobkey

DataServlet no longer uses ImageService to generate a URL for the image, and instead returns the BlobKey string of the image in its response.
script.js will pass this blobkey to fetchCommentImage to fetch the image, and generate a URL for it. This URL will then be used to display the image with the associated comment.
In this way, the ImageService dependency that does not work in production is removed.

* Change comment image width in display

* Rename url to imageUrl

url was assigned to the variable imageUrl before it was used anyway, so no need for the two names.

* Handle blob-key exceptional cases

When invalid blobkey is submitted to the /serve-images server, the server will respond with a 404 (image not found), which is handled in the front end by displaying "image not found" to the user, and console logging the error.
When no blobkey is provided, the server returns will return with null.