Change last timeslot of availability added to be timeslotStart
timeslotStart will be set to the end of the last event OR the start of the day if no events, so the last timeslot of availability will be either from the end of the last event until the end of the day OR  from the start of the day until the end of the day, when there are no events.
This avoids needing to retrieve the last event from the list of events, and also handles the edge case where there are no events.